
\name{des2}

\docType{data}

\alias{des2}

\title{Exemplary test design which may be typical in the context of large-scale assessment.}

\description{
This data set contains 39 test forms (booklets), each with 4 blocks of either reading, listening or orthography
tasks. Reading blocks start with "D-L", listening blocks start with "D-H", and orthography blocks start with
"D-R".
}

\usage{data(des2)}

\format{'data.frame':   39 obs. of  5 variables
  \describe{
    \item{TH}{booklet identifier}
    \item{X1}{first block}
    \item{X2}{second block}
    \item{X3}{third block}
    \item{X4}{fourth block}
 }
}

\source{artificial data}

\keyword{datasets}


