
\name{des3}

\docType{data}

\alias{des3}

\title{Exemplary test design for math and sciences (Biology, Chemistry, Physics).}

\description{
This data set contains 80 test forms (booklets), each with 6 blocks of either mathematics, biology, chemistry or
physics. Math blocks start with "Ma", biology blocks start with "Bio", chemstry blocks start with "Che", and
physics blocks start with "Phy".
}

\usage{data(des3)}

\format{'data.frame':   80 obs. of  10 variables
}

\source{artificial data}

\keyword{datasets}


