\name{trends}

\docType{data}

\alias{trends}

\title{Reading and listening achievement test data obtained from large-scale assessment at three times of measurement}

\description{
This data set contains fictional achievement scores of 7934 students in two domains (reading, listening) in the long format.
}

\usage{data(trends)}

\format{'data.frame':   291164 obs. of  11 variables:
  \describe{
    \item{year}{Year of evaluation}
    \item{idstud}{Student identifier}
    \item{idclass}{Class identifier}
    \item{country}{The country an examinee stems from}
    \item{language}{spoken language at home}
    \item{ses}{student's socio economical status}
    \item{sex}{student's sex}
  	\item{domain}{The domain the variable belongs to}
  	\item{task}{task identifier}
  	\item{item}{item identifier}
  	\item{value}{The response of the student to the item (0=incorrect; 1=correct)}
 }
}

\source{Simulated data}

\keyword{datasets}


